pyproject.toml
[project]
name = "codeshine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.2",
    "flask-swagger-ui>=5.21.0",
    "google-generativeai>=0.8.5",
    "loguru>=0.7.3",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "speechrecognition>=3.14.3",
]
